Reconfigured main recipe and added foreman extension
I'd like to be able to choose individual recipes, rather than include them all, but I think it's easier to abstract that out later.

+++ lib/tomafro/deploy/bundler.rb
@@ -1,7 +1,4 @@
-# Adds support for bundler within deployments.  This can be added to any `Capfile` by simply 
-# requiring it, e.g.:
-#
-#     require 'tomafro/deploy/bundler'
+# The bundler recipe ensures that the application bundle is installed whenever the code is updated.
 
 Capistrano::Configuration.instance(:must_exist).load do
   set(:bundle_gemfile) {"#{deploy_to}/Gemfile"}
@@ -10,20 +7,25 @@ Capistrano::Configuration.instance(:must_exist).load do
 
   namespace :bundle do
     namespace :install do
-      desc "Install the latest gem bundle"
-      task :default do
-        bundler "install --gemfile #{bundle_gemfile} --path #{bundle_dir} --deployment --quiet --without development test"
-      end
-
       desc "Install the latest gem bundle only if Gemfile.lock has changed"
       task :if_changed do
         if deployed_file_changed?(bundle_gemfile_lock)
           top.bundle.install
         end
       end
+
+      desc "Install the latest gem bundle"
+      task :default do
+        if deployed_file_exists?(bundle_gemfile)
+          bundler "install --gemfile #{bundle_gemfile} --path #{bundle_dir} --deployment --quiet --without development test"
+        else
+          puts "Skipping bundle:install as no Gemfile found"
+        end
+      end
     end
   end
 
-  # After deploy updates the code, install the latest bundle if required
+  # After deploy clones or updates the code, install the latest bundle (if needed)
+  after 'deploy:clone_code', 'bundle:install:if_changed'
   after 'deploy:update_code', 'bundle:install:if_changed'
 end
